requirements.txt
streamlit==1.32.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
chromadb==0.4.24
python-dotenv==1.0.1
pandas==2.2.1
tiktoken==0.6.0
ruff==0.3.4